changed folder/files structure
