+++ src/renderer/components/Dashboard/BimbeerCard.jsx
