+++ src/renderer/components/Dashboard/PictureCount.jsx
