+++ src/renderer/components/Setup/BeerList.jsx
