+++ src/renderer/components/Setup/SetupFormSteps/ProfileDiscoverySettingsForm.jsx
