+++ src/renderer/components/Setup/SetupFormSteps/ProfileFavBeerForm.jsx
