+++ src/renderer/components/Start/SignInForm.jsx
@@ -5,11 +5,11 @@ import { BsFacebook } from 'react-icons/bs';
 import { FcGoogle } from 'react-icons/fc';
 import { RiArrowRightLine } from 'react-icons/ri';
 import { getRedirectResult } from 'firebase/auth';
-import { UserAuth } from '../context/AuthContext';
-import { auth } from '../firebase/firebase';
-import logo from '../assets/images/logo.png';
+import logo from 'renderer/assets/images/logo.png';
+import { UserAuth } from '../../context/AuthContext';
+import { auth } from '../../firebase/firebase';
 import LoadingScreen from './LoadingScreen';
-import '../theme/css/AuthForms.css';
+import 'renderer/theme/css/AuthForms.css';
 
 export default function SignInForm() {
   const [email, setEmail] = useState('');
