+++ src/renderer/pages/Setup.jsx
@@ -1,4 +1,4 @@
-import SetupForm from 'renderer/components/SetupForm';
+import SetupForm from 'renderer/components/Setup/SetupForm';
 
 export default function Setup() {
   return <SetupForm />;
