+++ src/renderer/pages/SignUp.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignUpForm from '../components/SignUpForm';
+import SignUpForm from '../components/SignUp/SignUpForm';
 import ThemeBackground from '../components/ThemeBackground';
 import '../theme/css/AuthForms.css';
