+++ src/renderer/pages/Start.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignInForm from '../components/SignInForm';
+import SignInForm from '../components/Start/SignInForm';
 import '../theme/css/AuthForms.css';
 import ThemeBackground from '../components/ThemeBackground';
